(feat.) uiExtension/pointToPointTunnels: add editor popup skeleton
