+++ app/pages/ui/point-to-point-tunnels/point-to-point-tunnels-section.html
@@ -17,6 +17,7 @@
         sort-name="description"
         sort-reverse="false"
         is-sortable="false"
+        click-on-row="vm.editTunnel"
     >
     </ndm-table>
 
@@ -39,4 +40,5 @@
     </div>
 
     <br/><br/>
+    <!-- EDITOR -->
 </div>
